Advent of Code 2018, Day 21 Solution
Instead of directly tackling the 'code' solution
I tackled rewriting the code problem itself, reverse
engineering the program so I could figure out the
answer. The test.py is the variant that gives the
second solution (the first solution is just a matter
of running that and getting the first answer).

Unknown what the assembly 'input' looks like for
other people to know if the program can be adjusted
for others.

https://adventofcode.com/2018/day/21